requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Local vector db
faiss-cpu

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
streamlit
flask
pydantic

# Fetch transcript from youtube video
youtube-transcript-api

# Download audio for YouTube urls
yt_dlp

# Split downloaded audio files
# (such that we adhere to Whisper API's 25MB file size limit)
pydub
librosa
